net: Move the checksum functions to lib/
These functions are used by code outside the network support, so move them
to lib/ to be more accessible.

Without this, the functions are only accessible if CONFIG_NET is defined.
Many boards do not enable that option but still want to do checksums in
this format.

Fix up a few code-style nits while we are here.

Signed-off-by: Simon Glass <sjg@chromium.org>
Series-changes: 4
- Expand commit message to better explain the need to checksum functions