i2c: Tidy up designware PCI support
This is hacked into the driver at present. It seems better to have it as
a separate driver that uses the base driver. Create a new file and put
the X86 code into it.

Actually the Baytrail settings should really come from the device tree.

Note that 'has_max_speed' is added as well. This is currently always false
but since only Baytrail provides the config, it does not affect operation
for other devices.

Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Heiko Schocher <hs@denx.de>
Series-changes: 4
- Rename new file to designware_i2c_pci.c
- Add a comment in the commit message about why has_max_speed is added
- Add a comment about the speed logic in __dw_i2c_set_bus_speed()
- Drop unwanted debug printf("bad\n")
- Fix indentation nit