x86: coral: Enable TPM
Enable TPM2 so that we can use cr50.

Series-to: u-boot
Series-cc: bin, Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Cover-letter:
x86: coral: Add support for Cr50
This series adds a driver for the Cr50 security chip and enables it on
coral. This supports the 'tpm' command.

This series is built on the pending 'apollolake' series.
END

Signed-off-by: Simon Glass <sjg@chromium.org>