acpi: Convert part of acpi_table to use acpi_ctx
The current code uses an address where a pointer would result in fewer
casts. Also it repeats the alignment code which would be better done in a
helper function.

Update write_acpi_tables() to make use of the new acpi_ctx structure,
adding a few helps to clean things up.

Change-Id: I93e1e33891714417335e1dd517982b18bf9f882f
Signed-off-by: Simon Glass <sjg@chromium.org>