i2c: designware_i2c: Separate out the speed calculation
We want to be able to calculate the speed separately from actually setting
the speed, so we can generate the required ACPI tables. Split out the
calculation into its own function.

Drop the double underscore on __dw_i2c_set_bus_speed while we are here.
That is reserved for compiler internals.

Change-Id: I2e3d0961f7331632b382aaf38787ff8911e21149
Signed-off-by: Simon Glass <sjg@chromium.org>